+++ src/backports.c
@@ -37,32 +37,68 @@ the R sources.  Original copyright notices follow.
  *  along with this program; if not, a copy is available at
  *  https://www.R-project.org/Licenses/
  */
-
-/* 
- * Include after R header files 
+/*
+ * C level equivalent of `attributes()`, returns NULL or the attributes as a
+ * VECSXP.
  *
- * Code necessitated by the API restrictions implemented in R4.6+
+ * R_getAttributes was introduced in R 4.6.0. This backport uses ATTRIB which is
+ * still available in earlier R versions.
+ *
+ * The return value MUST BE PROTECTED.
+ *
+ * This code is from ~r89313
  */
 
+#include <Rinternals.h>
 #include <Rversion.h>
-#if R_VERSION < R_Version(4, 4, 1)
-#define allocLang Rf_allocLang
-
-static SEXP Rf_allocLang(int n)
-{
-    if (n > 0)
-          return LCONS(R_NilValue, Rf_allocList(n - 1));
-    else
-          return R_NilValue;
-}
-#endif
-
-#if R_VERSION < R_Version(4, 5, 0)
-# define R_ClosureFormals(x) FORMALS(x)
-# define R_ParentEnv(x) ENCLOS(x)
-
-#endif
 
 #if R_VERSION < R_Version(4, 6, 0)
-SEXP R_getAttributes(SEXP x);
+SEXP R_getAttributes(SEXP x)
+{
+    if (TYPEOF(x) == ENVSXP)
+	R_CheckStack(); /* in case attributes might lead to a cycle */
+
+    SEXP attrs = ATTRIB(x), namesattr;
+    int nvalues = length(attrs);
+    if (isList(x)) {
+	namesattr = getAttrib(x, R_NamesSymbol);
+	if (namesattr != R_NilValue)
+	    nvalues++;
+    } else
+	namesattr = R_NilValue;
+    /* FIXME */
+    if (nvalues <= 0)
+	return R_NilValue;
+    /* FIXME */
+    SEXP value, names;
+    PROTECT(namesattr);
+    PROTECT(value = allocVector(VECSXP, nvalues));
+    PROTECT(names = allocVector(STRSXP, nvalues));
+    nvalues = 0;
+    if (namesattr != R_NilValue) {
+	SET_VECTOR_ELT(value, nvalues, namesattr);
+	SET_STRING_ELT(names, nvalues, PRINTNAME(R_NamesSymbol));
+	nvalues++;
+    }
+    while (attrs != R_NilValue) {
+	SEXP tag = TAG(attrs);
+	if (TYPEOF(tag) == SYMSXP) {
+	    SET_VECTOR_ELT(value, nvalues, getAttrib(x, tag));
+	    SET_STRING_ELT(names, nvalues, PRINTNAME(tag));
+	}
+	else { // empty tag, hence name = ""
+            // Not reachable AFAIK
+            // nocov start
+	    MARK_NOT_MUTABLE(CAR(attrs));
+	    SET_VECTOR_ELT(value, nvalues, CAR(attrs));
+	    SET_STRING_ELT(names, nvalues, R_BlankString);
+            // nocov end
+	}
+	attrs = CDR(attrs);
+	nvalues++;
+    }
+    setAttrib(value, R_NamesSymbol, names);
+    UNPROTECT(3);
+    return value;
+}
 #endif