+++ src/fun.c
@@ -1,5 +1,5 @@
 /*
-Copyright (C) 2023 Brodie Gaslam
+Copyright (C) Brodie Gaslam
 
 This file is part of "vetr - Trust, but Verify"
 
@@ -166,7 +166,7 @@ struct ALIKEC_res ALIKEC_fun_alike_internal(
     }
   }
   UNPROTECT(6);
-  if(!res.success) res.wrap = allocVector(VECSXP, 2);
+  ALIKEC_res_wrap_check(&res);
   return res;
 }
 SEXP ALIKEC_fun_alike_ext(SEXP target, SEXP current) {