+++ src/lang.c
@@ -1,5 +1,5 @@
 /*
-Copyright (C) 2023 Brodie Gaslam
+Copyright (C) Brodie Gaslam
 
 This file is part of "vetr - Trust, but Verify"
 
@@ -247,9 +247,9 @@ struct ALIKEC_res ALIKEC_lang_obj_compare(
       res.dat.rec = ALIKEC_rec_ind_num(res.dat.rec, i + 2);
       res.dat.rec = ALIKEC_rec_dec(res.dat.rec);
     }
-    if(res.wrap == R_NilValue) res.wrap = allocVector(VECSXP, 2);
   }
   UNPROTECT(4);
+  ALIKEC_res_wrap_check(&res);
   return res;
 }
 
@@ -435,6 +435,7 @@ struct ALIKEC_res ALIKEC_lang_alike_rec(
     }
     res.dat.rec = ALIKEC_rec_dec(res.dat.rec);
   }
+  ALIKEC_res_wrap_check(&res);
   return res;
 }
 /*
@@ -527,15 +528,13 @@ SEXP ALIKEC_lang_alike_core(
     SEXP res_msg = PROTECT(allocVector(VECSXP, 2));
     SEXP res_msg_names = PROTECT(allocVector(STRSXP, 2));
     SET_VECTOR_ELT(res_msg, 0, ALIKEC_res_strings_to_SEXP(res.dat.strings));
-    if(res.wrap == R_NilValue) {
-      res.wrap = PROTECT(allocVector(VECSXP, 2));
-    } else PROTECT(R_NilValue);
+    // res.wrap already checked by ALIKEC_lang_alike_rec
     SET_VECTOR_ELT(res_msg, 1, res.wrap);
     SET_STRING_ELT(res_msg_names, 0, mkChar("message"));
     SET_STRING_ELT(res_msg_names, 1, mkChar("wrap"));
     setAttrib(res_msg, R_NamesSymbol, res_msg_names);
     SET_VECTOR_ELT(res_fin, 1, res_msg);
-    UNPROTECT(3);
+    UNPROTECT(2);
 
     SET_VECTOR_ELT(res_fin, 2, CAR(curr_cpy_par));
     SET_VECTOR_ELT(res_fin, 3, VECTOR_ELT(rec_ind, 0));
@@ -585,6 +584,7 @@ struct ALIKEC_res ALIKEC_lang_alike_internal(
     res.wrap = wrap;
   }
   UNPROTECT(2);
+  ALIKEC_res_wrap_check(&res);
   return res;
 }
 /*